requirements.txt
selenium==3.141.0
streamlit==1.29.0
webdriver-manager==4.0.1
beautifulsoup4==4.11.1
seleniumbase
webdriver-manager
scikit-surprise
pandas